// facturador-masmas/src/pages/Points/models/pointOfSale.ts
type pointOfSale = {
    name: string;
    province: string;
    department: string;
    locality: string;
    postalCode: string;
    street: string;
    number: number;
    height: number;
    email: string;
    phone: string;
    website: string;
    logo: string;
    color: string;
};
export default pointOfSale;